+++ src/deprecier-state.ts
@@ -1,0 +1,9 @@
+import { NpmConfig } from "./interfaces/npm.interface";
+import { PackageInfo } from "./interfaces/package-info.interface";
+import { RuleWithAppliedOptions } from "./interfaces/rule.interface";
+
+export class DeprecierState {
+  public rules: RuleWithAppliedOptions[] = [];
+  public packageInfo: PackageInfo | undefined = undefined;
+  public npmConfig: NpmConfig | undefined;
+}