+++ src/index.ts
@@ -8,22 +8,34 @@ import ConfigurationLoader from "./configuration-loader";
 import { deprecateAll } from "./rules/deprecate-all";
 import { supportLatest } from "./rules/support-latest";
 import { supportPreReleaseIfNotReleased } from "./rules/support-prerelease-if-not-released";
+import listActiveVersions from "./list-active-versions";
+import { DeprecierState } from "./deprecier-state";
+import fetch from "node-fetch";
+import execa from "execa";
+import { Npm } from "./npm";
+import { NpmApi } from "./npm-api";
+
+const npm = new Npm(execa);
 
 const oldVersionDeprecier = new OldVersionDeprecier(
-  new PackageInfoRetriever(),
+  new PackageInfoRetriever(new NpmApi(fetch), npm),
   new RuleApplier(new VersionForMessageFinder()),
-  new Deprecier(),
-  new Authentifier(),
+  new Deprecier(npm),
+  new Authentifier(npm),
   new ConfigurationLoader({
     deprecateAll: deprecateAll,
     supportLatest: supportLatest,
     supportPreReleaseIfNotReleased: supportPreReleaseIfNotReleased,
-  })
+  }),
+  listActiveVersions,
+  npm,
+  new DeprecierState()
 );
 
 module.exports = {
   verifyConditions: oldVersionDeprecier.verifyConditions.bind(
     oldVersionDeprecier
   ),
+  prepare: oldVersionDeprecier.prepare.bind(oldVersionDeprecier),
   publish: oldVersionDeprecier.publish.bind(oldVersionDeprecier),
 };