+++ src/old-version-deprecier.ts
@@ -1,7 +1,7 @@
 import { Config, Context } from "semantic-release";
-import { PluginConfig } from "./plugin-config";
+import { PluginConfig } from "./interfaces/plugin-config.interface";
 import { PackageInfoRetriever } from "./package-info-retriever";
-import { Action, RuleWithAppliedOptions } from "./rule";
+import { Action } from "./interfaces/rule.interface";
 import { RuleApplier } from "./rule-applier";
 import { SemVer } from "semver";
 import { Deprecier } from "./deprecier";
@@ -11,50 +11,73 @@ import {
 } from "./rule-application-result";
 import { Authentifier } from "./authentifier";
 import ConfigurationLoader from "./configuration-loader";
-import { PackageInfo } from "./package-info";
+import { PackageInfo } from "./interfaces/package-info.interface";
+import { ListActiveVersions } from "./list-active-versions";
+import { DeprecierState } from "./deprecier-state";
+import { Npm } from "./npm";
 
 export class OldVersionDeprecier {
-  private rules: RuleWithAppliedOptions[] = [];
-
   constructor(
     private readonly packageInfoRetriever: PackageInfoRetriever,
     private readonly ruleApplier: RuleApplier,
     private readonly deprecier: Deprecier,
     private readonly authentifier: Authentifier,
-    private readonly configurationLoader: ConfigurationLoader
-  ) {
-    this.packageInfoRetriever = packageInfoRetriever;
-    this.ruleApplier = ruleApplier;
-    this.deprecier = deprecier;
-    this.authentifier = authentifier;
-  }
+    private readonly configurationLoader: ConfigurationLoader,
+    private readonly listActiveVersions: ListActiveVersions,
+    private readonly npm: Npm,
+    private readonly deprecierState: DeprecierState
+  ) {}
 
   public async verifyConditions(pluginConfig: PluginConfig): Promise<void> {
-    this.rules = this.configurationLoader.generateRules(pluginConfig);
+    this.deprecierState.rules = this.configurationLoader.generateRules(
+      pluginConfig
+    );
+  }
+
+  public async prepare(
+    pluginConfig: PluginConfig,
+    context: Context & Config
+  ): Promise<void> {
+    this.deprecierState.npmConfig = await this.npm.getConfig(context);
+    this.deprecierState.packageInfo = await this.packageInfoRetriever.getInfo(
+      this.deprecierState.npmConfig,
+      context
+    );
   }
 
   public async publish(
     pluginConfig: PluginConfig,
     context: Context & Config
   ): Promise<void> {
-    const { logger, cwd, env } = context;
-    const packageInfo = await this.packageInfoRetriever.getInfo({
-      cwd,
-      env,
-      logger,
-    });
-
-    if (!packageInfo) {
+    const { logger } = context;
+    if (!this.deprecierState.packageInfo) {
+      logger.log("This project does not seem to be a npm package");
       return;
     }
 
-    const parsedVersions = packageInfo.versions.map((v) => new SemVer(v));
-    const actionsOnVersions = this.ruleApplier.applyRules(
-      parsedVersions,
-      this.rules
+    if (!this.deprecierState.npmConfig) {
+      throw new Error(
+        "Unable to deprecate version as the configuration of NPM could not be retrieved"
+      );
+    }
+
+    const activeVersions = this.listActiveVersions(
+      this.deprecierState.packageInfo
     );
 
-    await this.deprecate(actionsOnVersions, packageInfo, context);
+    logger.log(`Active versions: ${activeVersions.join(", ")}`);
+
+    const parsedVersions = activeVersions.map((v) => new SemVer(v));
+    const actionsOnVersions = this.ruleApplier.applyRules(
+      parsedVersions,
+      this.deprecierState.rules
+    );
+
+    await this.deprecate(
+      actionsOnVersions,
+      this.deprecierState.packageInfo,
+      context
+    );
   }
 
   private async deprecate(
@@ -66,13 +89,16 @@ export class OldVersionDeprecier {
       .filter((actionOnVersion) => actionOnVersion.action === Action.deprecate)
       .map((actionsOnVersion) => actionsOnVersion as DepreciationResult);
 
-    if (depreciations) {
+    if (depreciations?.length) {
       context.logger.log(
         "Versions to deprecate",
         ...depreciations.map((v) => v.version.format())
       );
 
-      await this.authentifier.authentify(context);
+      await this.authentifier.authenticate(
+        this.deprecierState.npmConfig!,
+        context
+      );
       for (const depreciation of depreciations) {
         await this.deprecier.deprecate(packageInfo, depreciation, context);
       }